Who do We Support Page: I would like to keep the color bars all the same size and have the type be larger to fill the area as in the layout. That way each section will be equal in size.
How Do We Do This Page:
I would like to keep the color bars all the same size and have the type be larger to fill the area as in the layout. That way each section will be equal in size.

Bios Pages:
The Photos seem too large. Maybe reduce 25% I think the copy can also be changed to Gotham light. It feels heavy. On Both.
Titles on both photos can be closer to the image.
Contact Us Page: Needs copy and design to be provided. Should this appear on the home page as address and phone numbers or on bios page?
Carousel multi images need to be provided.
Mobile/Responsive Home Page:
The logo looks too small. Can we increase the size here.